+++ pkg/registry/distribution_test.go
@@ -7,6 +7,42 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+func TestReference(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name     string
+		refName  string
+		expected bool
+	}{
+		{
+			name:     "with latest tag",
+			refName:  "ghcr.io/spegel-org/spegel:latest",
+			expected: true,
+		},
+		{
+			name:     "no latest tag",
+			refName:  "ghcr.io/spegel-org/spegel:v1.0.0",
+			expected: false,
+		},
+		{
+			name:     "empty name",
+			refName:  "",
+			expected: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			ref := reference{
+				name: tt.refName,
+			}
+			require.Equal(t, tt.expected, ref.hasLatestTag())
+		})
+	}
+}
+
 func TestParsePathComponents(t *testing.T) {
 	t.Parallel()
 
@@ -24,6 +60,14 @@ func TestParsePathComponents(t *testing.T) {
 			path:            "/v2/foo/bar/manifests/hello-world",
 			expectedName:    "example.com/foo/bar:hello-world",
 			expectedDgst:    "",
+			expectedRefKind: referenceKindManifest,
+		},
+		{
+			name:            "valid manifest digest",
+			registry:        "docker.io",
+			path:            "/v2/library/nginx/manifests/sha256:0a404ca8e119d061cdb2dceee824c914cdc69b31bc7b5956ef5a520436a80d39",
+			expectedName:    "",
+			expectedDgst:    digest.Digest("sha256:0a404ca8e119d061cdb2dceee824c914cdc69b31bc7b5956ef5a520436a80d39"),
 			expectedRefKind: referenceKindManifest,
 		},
 		{