+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [#717](https://github.com/spegel-org/spegel/pull/717) Extend tests for distribution.
+
 ### Changed
 
 ### Deprecated